requirements.txt
# Requirements lower bounds listed here are our best effort to keep them up to
# date but we do not test them so no guarantee of having them all correct. If
# you find any incorrect lower bounds, let us know or propose a fix.

pbr>=6.0.0 # Apache-2.0
keystonemiddleware>=5.1.0 # Apache-2.0
lxml>=2.3 # BSD
oslo.config>=6.8.0 # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.log>=4.3.0 # Apache-2.0
oslo.policy>=4.5.0 # Apache-2.0
PasteDeploy>=1.5.0 # MIT
Paste>=3.10 # MIT
pecan>=0.8.0 # BSD
oslo.middleware>=3.22.0 # Apache-2.0
oslo.utils>=4.7.0 # Apache-2.0
WSME>=0.12.1 # MIT
python-observabilityclient>=0.0.4 # Apache-2.0
